refactor: 응답 직렬화 위치 이동
