+++ tomcat/src/main/java/org/apache/coyote/http11/HttpResponse.java
@@ -1,5 +1,11 @@
 package org.apache.coyote.http11;
 
+import java.io.File;
+import java.io.IOException;
+import java.net.URISyntaxException;
+import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -32,16 +38,10 @@ public class HttpResponse {
         return String.join("\r\n", headerString);
     }
 
-    public HttpStatusCode getStatusCode() {
-        return statusCode;
-    }
-
-    public String getContentType() {
-        return contentType.get();
-    }
-
-    public String getBody() {
-        return body;
+    public String createString() throws IOException, URISyntaxException {
+        String headerString = headersToString();
+        String responseBody = bodyToString();
+        return buildResponse(headerString, responseBody).toString();
     }
 
     public void addCookie(String key, String value) {
@@ -50,5 +50,33 @@ public class HttpResponse {
 
     public void setLocation(String value) {
         headers.put("Location", value);
+    }
+
+    private StringBuilder buildResponse(String headerString, String responseBody) {
+        StringBuilder responseBuilder = new StringBuilder();
+        responseBuilder.append("HTTP/1.1 ")
+                .append(statusCode.getValue()).append(" ")
+                .append(statusCode).append(" \r\n");
+        if (headerString != null) {
+            responseBuilder.append(headerString).append(" \r\n");
+        }
+        responseBuilder.append("Content-Type: ").append(contentType.get()).append(" \r\n");
+        responseBuilder.append("Content-Length: ").append(responseBody.getBytes().length).append(" \r\n\r\n");
+        responseBuilder.append(responseBody);
+        return responseBuilder;
+    }
+
+    private String bodyToString() throws IOException, URISyntaxException {
+        URL resource = getResource(body);
+        if (resource == null || Files.isDirectory(Path.of(resource.toURI()))) {
+            return "Hello world!";
+        }
+        return new String(Files.readAllBytes(new File(resource.getFile()).toPath()));
+    }
+
+    private URL getResource(String path) {
+        return getClass()
+                .getClassLoader()
+                .getResource("static" + path);
     }
 }
